requirements.txt
# Project Dependencies
google-generativeai
tweepy
requests
python-dotenv